requirements.txt
qwen-tts==0.1.1
transformers
accelerate
torch
torchaudio
torchvision
soundfile
librosa
tqdm
opencv-python
numpy
gradio
google-genai
